[bash scripts] Switch from dashes to underscores for function names.
Underscores conform to the official spec for function names[1]. Dashes
mostly work, but fail in some cases, such as when trying to export them.

Further, function names with underscores are treated as a single word
in vim and so searching for them with * or # works. If searching that
way for functions with dashes in their names, only the segment of the
name with the cursor on it is searched for.

[1] http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_09_05