[init.vim] Set 'mouse' option
Set the 'mouse' option to 'a', which enables it for all (normal, visual,
insert, and command) modes. This allows users to move the cursor, select
text by dragging, select text by multi-clicking, select windows, and
resize windows.

It can also be used to scroll the buffer (by using the scroll wheel or
equivalent). But the implementation seems to be clunky. Scrolling is too
quick and jerky. I'd like to slow it down a little. But I can't find any
options to control that.

Still, set this option for now and see if I like it.