[zshrc] zsh-ify is_executable
is_executable is a shell function that determines if its argument is
something that can be executed, be it a tool, alias, or function.
is_executable was originally written to use bash's `command -v`
function. However, this function operates a little differently under
zsh. `command -v` is equivalent to `whence`, while what we actually want
is the behavior of `whence -p`. So change is_executable to use that
instead.