[zshrc, install.brew] Enhance brew_path
The brew_path function is used to return the base path of our homebrew
installation, if any. Previously, it was implemented in terms of the
brew command line tool, invoking `brew --prefix`. However, `brew_path`
is sometimes used in a chicken-and-egg situation. In particular, we use
it to configure $PATH to include the homebrew installation. Naturally,
we can't invoke `brew` in order to find out where homebrew is so that we
can add it to the path so that we can invoke `brew` in order to find out
where homebrew is...

So, enhance brew_path. If `brew` is available, use it. Otherwise, use
some hard-coded knowledge of where it is typically installed in order to
try to find it.