[zshrc] Rework the 'help' alias
Support for the 'help' alias was kind of sparse. I found some lines on
the internet some time ago that loaded run-helper and aliased run to
invoke it. But there were no comments about those lines. So I've added
some comments, cleaned up the code a little, and left behind a
commented-out alternative that tries to look up command help in
zshbuiltins.