[zshrc] Refactor v()
I was using v() as the main implementation for invoking my preferred
editor, and then mapping vi, vim, and nvim as aliases for it. However,
while I like the idea of using v, vi, vim, nvim interchangeably, I
didn't like the asymmetry with v having the implementation and the
others being aliases for it. Therefore, move the implementation into
vim_core() and have v, vi, vim, and nvim invoke *that*.