[README, install, zshenv, zshrc] Update history
As part of cleaning up my home directory, arrange to have zsh history
data written to ~/.config/zsh. Along with this, disable the facility in
/etc/zshrc_Apple_Terminal that manages per-session history and, in turn,
which creates the ~/.zsh_sessions directory. Note that this might break
the recovery of sessions that were abruptly aborted, but I think I can
live with that.

Add a hist() shell function that maps to `history -i 0` -- that is, show
all history along with full ISO timestamps.

While we're at it, also move ~/.zcompdump to ~/.config/zsh.