[realpath] Reimplemented in terms of `readlink -f`
When I first wrote this script/function, it was because the version of
readlink that came with macOS didn't support the -f option[1], which
-- in the Linux version -- would "canonicalize by following every
symlink in every component of the given name recursively". Therefore,
Mac users needed to write a script that would perform that
canonicalization itself component by component.

The version of readlink in macOS 12.3[1] has since received the -f option.
This option is is the same as for the stat command (stat and readlink
are the same binary), though slightly different than the Linux version
-- the BSD version is used to provide a format specification -- but that
option can be used to achieve the same effect as the Linux flag.
Therefore, update and simplify this script/function to use that option.

[1] macOS 12.3 was released in March 2022. It looks like the option was
added to the FreeBSD sources in December 2010[2], which was included in
FreeBSD 9.0 in Jan 2012[3].

[2] https://cgit.freebsd.org/src/commit/usr.bin/stat/stat.c?id=b9f9338ea87a2ad5e901e6bc61fdbef06a25b004
[3] https://cgit.freebsd.org/src/tag/?h=release/9.0.0