[init.vim] Disable some Python-formatting warnings
Disable the warnings related to lines being split before or after a
binary operator. Both of these were enabled, making it impossible to
split a line in the vicinity of a binary operator (which is a common
place to want to split a line). I've never settled on whether it's
better to split a line before a binary operator or after, so I'm
disabling both warnings.