[install.brew] Add python-lsp-server modules
The python-lsp-server github project suggests a number of python modules
to install for maximum goodness. Update install.brew to install them
after python3 is installed.

While doing this, added a facility for checking what Python modules are
already installed and use that information to only attempt to install
new modules. Generalize this and also use it when installing brew
formulae and casks.