[init.vim] Change mapping for AutoFormat
I used to use <leader>ff for invoking AutoFormat. However, I am now also
trying to use <leader>f for something else. Because of the overlap
between the starts of these two sequences, [n]vim needs to pause a
little bit to see if the second "f" is coming. To avoid that pause,
change <leader>ff to <leader>af.