[init.vim] Fiddle with 'highlight' colors
Try to stick as close to possible with the defaults. Use the 'default'
color scheme, and remove the apparently unneeded 'background' setting.

Set 'CursorLine cterm=none', which remove the underlining of the line
that contains the cursor. On Help pages, at least, moving up in a
document will leave a trail of underlined lines if this is left on.

Set the ColorColumn to a nice light grey.