[init.vim] Remove leftover code completion support
Remove some code leftover from an earlier attempt to get code completion
working. This code used a plugin that I want to stay away from for now
since its main purpose in life is to provide a unified async interface
for both vim and nvim. I want to commit to nvim, and so don't want to
rely on a compatibility interface if I don't have to.