[init.vim] Remap 'n' and 'N'.
Remap 'n' and 'N' to search for the next match (as is standard) but to
also center the found text vertically. This is better than the standard
approach, which is to just highlight the next match if it is visible
later in the window, or to scroll the next match to the center location
if it was previously outside the window.