[init.vim] Fiddle with auto-completion menu appearance
nvim-cmp has an experimental.native_menu setting to control how the
popup menu appears. This was set to 'false' because the .vim file I
copied it from had it set that way. I tried setting it to 'true', but
was told that that option was now deprecated and that I should try
setting view.entries to custom, native, or wildmenu. I didn't like
wildmenu, so I'm not using that one. I didn't see any difference between
custom and native, so I'll just leave it at native for now.