[init.vim] Update rust-tools configuration
When using rust-tools, its documentation says to let it set up LSP
configuration and rust-analyzer integration. This means that we need to
disable the code that we already had that did that. But in doing so, we
also removed the code that set up our keybindings. Restore those by
moving them to the rust-tools configuration section.