[init.vim] Replace custom treesitter parser installation code
When setting up treesitter support, I added some lua code to take a list
of desired parsers, compare it to the set of installed parsers, and
perform installations and uninstallations as needed. However, just after
I finished all of that (the process of which was a quite enjoyable
learning experience), I stumbled across the nvim-treesitter
documentation that showed how I could do pretty much the same thing by
calling its setup() routine. It won't do the uninstallations, but I'm OK
with that. So...delete all my old code and replace it with a simple call
to setup().