[update] Remove
This script was mostly to help update plugins for both vim and nvim, as
well as homebrew. I now have moved completely over to nvim, so I don't
need to support vim. And I'll let packer manage updating the plugins
as-needed. This left the script only needing to update homebrew, which I
can do by hand without a script.