[install.brew, init-autoformat] Use Python 3.11
Invoking Python with `python` will invoke the default system Python, be
it Python 2 or 3. We generally want to use Python3, which is typically
invoked with `python3`. However, python3 is linked to Python 3.9, which
is my system-installed Python. I want Python 3.11, which is my
Homebrew-installed Python. Installing Python 3.11 currently doesn't set
up links for python3, pip3, etc., so I need to explicitly specify that I
want python3.11 in order to invoke it.