[misc.lua] Compatibility with latest nvim
There seems to be a problem with using local variables in a script
passed to vim.cmd. They don't seem to be supported, and we get an error
message when trying to use them. Avoid this for now by not saving and
restoring the original cursor position when running our StripWhitespace
utility.