[plugins.lua] Add comment about configuring mccabe analyzer
It's all too easy to trigger warnings about exceeding the mccabe
complexity cyclomatic threshold. I'm not triggering it right now in a
python project I'm working on, but I did earlier and I may do so again.
It took a while to figure out how to configure the threshold value, so
I'm checking in a comment on how to do that in the future.