[*.lua] Move files from lua/ to lua/user/.
This was a suggestion from chris@machine of LunarVim. The idea here is
to create a "namespace" for your custom configuration files. The need
for a namespace became apparent to me when I tried to create a file
called "lazy.lua" to load and configure the "lazy.nvim" plugin. Both of
these would be accessed with "require('lazy')", leading to ambiguity. To
resolve that, I used the name "plugins.lua" instead of "lazy.lua". But
this next led to problems of where to store my Lua files that
initialized my plugins. I wanted to put them in a directory called
"plugins". But this would in turn have a conflict with "plugins.lua" --
"require('plugins')" wouldn't know which one to access. So I called the
directory "init-plugins". I didn't like either of these names, so when I
saw chris's suggestion, I learned that it would solve my naming problem.
So:

* Moving my files from lua/ to lua/user/
* Renaming plugins.lua to lazy.lua
* Renaming init-plugins to plugins