[events.lua] Fix restoration of cursor when opening a file
I have a little autocommand that gets called when a file is opened that
restores the file's previous cursor position (rather than letting it
always reset to the top of the file). Further, this autocommand would
disable itself for git commit and rebase buffers, since it's best to
start at the tops of those files. Maybe it never worked, maybe it only
later stopped working, but it seems that triggering this code on
BufReadPost was/is not the right thing to do. Checking the file type at
that time for 'gitcommit' or 'gitrebase' doesn't work, since the file
type is not yet established. Instead, trigger on the FileType event.