[misc.lua] Remove unused funtions
Remove Reload function. This function worked well when everything was in
a single vimrc file, but no longer works with the massive suite of lua
files I now have, especially with all those plugins. It's now much
simpler to just restart vim.

Remove WhatAmI function. I don't remember exactly why I wanted/needed
it. But it seems silly to have given that the name of the file I'm
working on is there in the status line.