[init.lua, lazy.lua, prelazy.lua] Move to new method for plugins
Use the approach suggested by chris@machine, where specs for loading
plugins via lazy.nvim are specific explicitly. Having them loaded
implicitly is nice, but it's a little hidden. As well, being able to
open init.lua and using it as a "table of contents" which can be used as
a launching point to loading all the other .lua files seems nice.