[several plugin files] Replace commentary with Comment.nvim
Comment.nvim seems to have some nice features (like full motion support
and things like gco, gcO, and gcA) and works with treesitter for more
accurate results.

Note that I had to tweak which-key's timeout value in order to get gcO
and gcA to work. The issue seems to be related to:

https://github.com/numToStr/Comment.nvim/issues/115#issuecomment-1032290098