[lsp.lua] Fiddle with how LSP plugins are managed
I realized that mason-tool-installer could install everything I need, so
I relieved mason-lspconfig of that duty. I also removed the custom icons
used in the Mason dialog, simplified some in-dialog keymappings and
rewrote the explanatory text at the top of the file to reflect the
change in how tools are loaded.