[lsp.lua] Remove clangd as an LSP server
Out of the box, clangd will report too many false issues. It doesn't
know, for example, how to load the includes of a large project. Without
being able to do that, it doesn't know about external structs and other
types, functions, macros, etc. The noise is deafening. So remove clangd
to until I can get around to figuring out how to soothe it.