[which-key.lua] Change mapping for nohlsearch
I used to have this set to 'l'. Then I decided to use 'l' as a prefix
for LSP-related mappings, so I moved it to 'h'. But I'm finding that too
awkward, so I'm now setting it to 'll'. It "pollutes" the 'l'
prefix-space, but I can live with that.