[which-key.lua] Stop ZZ from whining
I like to use ZZ for quitting vim. I've always used it, for so long that
there's no way I can stop using it. But I hate how it warns that I
haven't necessarily visited every file that I specified on the command
line. It warns that there are "X more files to edit". I don't care. I
just want to quit. So let's map "ZZ" to ":xa", which will do the same
thing without whining. I think. The help pages on the various :q, :cq,
:wq, and :x commands -- optionally with "a" and/or "!" suffixes -- are
confusing and possibly wrong. But it seems to be working, so I'll try
it.