[zshrc] Don't open dunder files with 'vap'
The 'vap' command is used to edit all Python files in the current
directory. This is handy when working on a small set of related files.
However, when the directory is configured to be part of a package, there
may be some dunder files (__about__.py, __init__.py, __main__.py) in it.
I almost never want to edit those, so make vap ignore them.