[zshrc] Update `py` to know about venv's
`py` now searches for python as follows:

* Search the directory tree upwards for a python virtual environment, as
  identified by the existence of a pyvenv.cfg file. If one is found, use
  the python that's in its 'bin' directory, if it exists.
* Use the homebrew version, if it exists.
* Use the system-installed version, if it exists.
* Only look for instances of python3. Don't look for python3.12,
  python3.11, or just plain old python.

In this way, we can automatically use a virtual environment by using the
`py` command -- there's no need to activate/deactivate. If we're not in
a virtual environment, we'll fallback to global installations. If we're
in a virtual environment but want to use the global installation, then
we can invoke python with `python3`.