[install, zshrc] Simplify getting the current directory
In both install and zshrc, I need to get the current directory. I need
the current directory in install in order to get access to resources in
the same directory (such as the dotfiles I'm installing into $HOME). I
need it in zshrc for my cdd() command and so I can add my private bin
directory to $PATH. In the former case, things are pretty straightfoward
-- I can work from the path in $0. However, getting the name/path of the
zshrc file is tricker. There's a succinct solution for zsh, but it's
pretty inscrutible, so I added some comments explaining it.