[events.lua] Protect cursor restoration
When restoring the cursor position when opening a file, make sure we
don't try to set the positon past the end of the file (which can happen
if the file is shortened outside of vim after vim last edited it). If we
were to try to do that, we'd get a big ass error message, buried within
it being "Cursor position outside buffer".