[install,install.python_packages] Remove install.python_packages
The Homebrew install of python 3.12 no longer lets you simply install
packages into its site-packages directory. Perhaps there's a way to
override that, but the recommended approach is to now use virtual
environments for everything. The only package I was installing was
Beautiful Soup for the APOD script, so I've converted that project to
use a venv. With that done, I no longer need install.python_packages.