[lsp] Add back beautysh. Sort of.
The beautysh shell formatting tool has been deprecated from none-ls.
However, because neither of the suggested replacements for it
(bashls/bash-language-server or shfmt) support zsh, beautysh has been
added to the null-ls-extras repository. But it's not apparent how to
access these extra tools. I've made my best guess, but it's not working.
I'll check in my (commented-out) work for now, hoping that the shape of
the new repo (it's actually only a few minutes old) is updated.