[install.brew] Install ruff and uv
These are rust-based python tools I'm interested in trying out. They're
supposed to replace (in the case of ruff) a host of python formatting
and linting tools and (in the case of uv) pip, pip-tools, venv, and even
more over time. Their main attractions are that they're fast and that
they consolidate lots of other tools under one roof, allowing them
opportunities for some nice optimzations.