[shell_functions] Attempt to speed up find_python_root
The original version of find_python_root would start in the current
directory, look around for pyvenv.cfg files, and the move up the
directory hierarchy if it didn't find one, ultimately stopping at the
root ("/"). That approach is retained, but before we try it, we first
look at $VIRTUAL_ENV. If it exists and it looks like the current
directory is under/in it, then use that as the python project root. This
way, we are doing a simple string match rather than performing a lot of
file system operations.