[zshrc] Make displaying the prompt less annoying
The prompt is configured to display virtual environment and git
information. Both states are checked every time the prompt is displayed,
but performing these checks are annoying: they result in commands being
executed and displayed in my terminal tab. It turns out that these
commands are triggered if GIT_PS1_SHOWUNTRACKEDFILES is set. I can live
without that, so remove it.