[install, shell_functions] Take a more-considered approach to HERE and DOTFILES
HERE and DOTFILES were generally considered to be the same thing. And --
for now -- they are. But I'm thinking that the install scripts -- the
location of which is identified by HERE -- might be moved to a
subdirectory in order to keep them out of the way. In that case, HERE
would refer to that subdirectory, and DOTFILES would continue to refer
to the root of the repository location. The could no longer be used
interchangeably. Make changes with that possibility in mind.